// crypto-vault-expo/src/features/home/NFTCollection/ton/NFTImport/NFTTonImport.type.ts
export type NFTTonResponse = {
    name: string;
    owner: string;
    symbol: string;
    tokenURI?: string;
    tokenStandard?: NFTTokenStandard;
    quantity?: number;
};
// ...
export type NFTTonRootData = Omit<NFTTonResponse, 'quantity'> & {
    contractAddress: string;
    id: string;
    accountId: string;
    protocol: ProtocolDataFromBEType;
    tonCollectionAddress: string;
};

export type NFTTonType = {
    description?: string;
    external_url?: string;
    image?: string;
    name?: string;
    attributes?: [];
    network_image?: string;
    image_data?: string;
    quantity?: number;
    tokenStandard?: NFTTokenStandard;
    nftDetailAll: NFTDetailEVMCollectionType & Nftitem;
};

export type NFTTonData = {
    root: NFTTonRootData;
    detail: NFTTonType;
};
// ...
export interface NFTDetailEVMCollectionType {
    token_address: string;
    token_id: string;
    amount: string;
    token_hash: string;
    block_number: string;
    block_number_minted: null | string;
    contract_type: string;
    name: string;
    symbol: string;
    token_uri: string;
    metadata: string | Record<string, any> | null;
    normalized_metadata?: Record<string, any> | null;
    media?: Array<{
        mimetype?: string;
        parent_hash?: string;
        status?: string;
        updatedAt?: string;
        media_collection?: {
            low?: { url?: string };
            high?: { url?: string };
        };
        original_media_url?: string;
    }>;
    last_token_uri_sync: string;
    last_metadata_sync: string;
    minter_address: null | string;
    owner_of: string;
    rarity_rank: number;
    rarity_percentage: number;
    rarity_label: string;
    possible_spam: boolean;
    verified_collection: boolean;
    floor_price: null | string;
    floor_price_usd: null | string;
    floor_price_currency: null | string;
    active: boolean;
}